pyproject.toml
[tool.poetry]
name = "INSTA_BOT"
version = "0.1.0"
description = ""
authors = ["FatFry"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.17.2"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
